# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: XAGUSD Screener API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: signals
    description: XAGUSD signal operations
  - name: push
    description: Web Push subscription management for browser notifications
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /levels:
    get:
      operationId: getLevels
      tags: [signals]
      summary: Get key price levels and clear trade signal
      description: Returns pivot-point-derived support/resistance levels, buy/sell zones, and a single clear trade signal for the requested symbol
      parameters:
        - in: query
          name: symbol
          schema:
            type: string
            enum: ["XAGUSD", "XAUUSD", "EURUSD", "GBPUSD", "AUDUSD", "USDJPY", "GBPJPY", "BTCUSD", "ETHUSD"]
            default: "XAGUSD"
          description: Trading instrument
        - in: query
          name: timeframe
          schema:
            type: string
            enum: ["15m", "30m", "1h", "1d"]
            default: "1d"
          description: Bar timeframe used to compute pivots and zones
        - in: query
          name: accountSize
          schema:
            type: number
            minimum: 1
            default: 500
          description: Trading account size in USD used to compute position sizing
        - in: query
          name: riskPct
          schema:
            type: number
            minimum: 0.01
            maximum: 100
            default: 1
          description: Percent of account risked per trade (1 = 1%)
        - in: query
          name: minCollateral
          schema:
            type: number
            minimum: 0.01
            default: 10
          description: Minimum collateral the exchange will accept per position (USD). Defaults to Jupiter perps minimum of $10.
        - in: query
          name: maxLeverage
          schema:
            type: number
            minimum: 1
            maximum: 200
            default: 50
          description: Maximum leverage the trader is willing to use. Defaults to 50x for Jupiter perps.
        - in: query
          name: mt5Lots
          schema:
            type: number
            minimum: 0.01
            maximum: 100
            default: 0.01
          description: MetaTrader 5 lot size used to size forex/metals positions. Ignored for crypto perps (BTC/ETH) which use Jupiter collateral × leverage. Default 0.01 (1 micro lot).
      responses:
        "200":
          description: Price levels and signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LevelsData"

  /backtest:
    get:
      operationId: getBacktest
      tags: [signals]
      summary: Backtest the signal criteria over historical candles
      description: Replays the buy/sell zone signal logic over historical candles for the requested symbol and timeframe.
      parameters:
        - in: query
          name: symbol
          schema:
            type: string
            enum: ["XAGUSD", "XAUUSD", "EURUSD", "GBPUSD", "AUDUSD", "USDJPY", "GBPJPY", "BTCUSD", "ETHUSD"]
            default: "XAGUSD"
          description: Trading instrument
        - in: query
          name: timeframe
          schema:
            type: string
            enum: ["15m", "30m", "1h", "1d"]
            default: "1d"
          description: Bar timeframe used for the backtest
      responses:
        "200":
          description: Backtest results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestResult"

  /active-signals:
    get:
      operationId: getActiveSignals
      tags: [signals]
      summary: All currently-active BUY/SELL signals across symbols × timeframes
      description: Recomputes every tracked symbol × timeframe and returns the BUY/SELL signals as a single overview. Each entry includes the timeframe alongside the standard LevelsData payload so the client can render and deep-link without extra calls. Sizing inputs (accountSize, riskPct, minCollateral, maxLeverage) flow through to every entry's positionSizing block — pass the trader's actual settings so the overview's $P&L matches the main signal panel.
      parameters:
        - in: query
          name: accountSize
          schema:
            type: number
            minimum: 1
            default: 500
          description: Trading account size in USD used to compute position sizing for every entry
        - in: query
          name: riskPct
          schema:
            type: number
            minimum: 0.01
            maximum: 100
            default: 1
          description: Percent of account risked per trade (1 = 1%)
        - in: query
          name: minCollateral
          schema:
            type: number
            minimum: 0.01
            default: 10
          description: Minimum collateral the exchange will accept per position (USD). Defaults to Jupiter perps minimum of $10.
        - in: query
          name: maxLeverage
          schema:
            type: number
            minimum: 1
            maximum: 200
            default: 50
          description: Maximum leverage the trader is willing to use. Defaults to 50x for Jupiter perps.
        - in: query
          name: mt5Lots
          schema:
            type: number
            minimum: 0.01
            maximum: 100
            default: 0.01
          description: MetaTrader 5 lot size used to size forex/metals positions in every entry. Ignored for crypto perps (BTC/ETH).
      responses:
        "200":
          description: Active signals overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActiveSignalsResponse"

  /push/vapid-public-key:
    get:
      operationId: getVapidPublicKey
      tags: [push]
      summary: Get the server's VAPID public key for browser subscription
      description: Returns the public half of the VAPID keypair used to sign Web Push payloads. The browser passes this to PushManager.subscribe() before posting the resulting subscription back via /push/subscribe.
      responses:
        "200":
          description: VAPID public key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VapidPublicKey"
        "503":
          description: VAPID keys not configured on the server

  /push/subscribe:
    post:
      operationId: subscribePush
      tags: [push]
      summary: Register a browser Web Push subscription
      description: Stores a Web Push subscription so the server can fan out signal alerts to it. Idempotent — re-posting the same endpoint upserts the existing record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscriptionPayload"
      responses:
        "200":
          description: Subscription stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushSubscribeResponse"

  /push/unsubscribe:
    post:
      operationId: unsubscribePush
      tags: [push]
      summary: Remove a browser Web Push subscription
      description: Deletes the subscription matching the supplied endpoint. Safe to call when no matching row exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushUnsubscribeRequest"
      responses:
        "200":
          description: Subscription removed (or already absent)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushSubscribeResponse"

  /price-history:
    get:
      operationId: getPriceHistory
      tags: [signals]
      summary: Get price history aligned to live OANDA spot
      parameters:
        - in: query
          name: symbol
          schema:
            type: string
            enum: ["XAGUSD", "XAUUSD", "EURUSD", "GBPUSD", "AUDUSD", "USDJPY", "GBPJPY", "BTCUSD", "ETHUSD"]
            default: "XAGUSD"
          description: Trading instrument
        - in: query
          name: timeframe
          schema:
            type: string
            enum: ["15m", "30m", "1h", "1d"]
            default: "1d"
          description: Bar timeframe
        - in: query
          name: bars
          schema:
            type: integer
            default: 200
          description: Number of bars to return
      responses:
        "200":
          description: Price history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceHistory"

components:
  schemas:
    VapidPublicKey:
      type: object
      properties:
        publicKey:
          type: string
          description: Base64URL-encoded VAPID public key suitable for PushManager.subscribe applicationServerKey
      required:
        - publicKey

    PushSubscriptionKeys:
      type: object
      properties:
        p256dh:
          type: string
          description: Browser-supplied P-256 ECDH public key (base64url)
        auth:
          type: string
          description: Browser-supplied auth secret (base64url)
      required:
        - p256dh
        - auth

    PushSubscriptionPayload:
      type: object
      properties:
        endpoint:
          type: string
          description: Push service endpoint URL (FCM / APNs gateway)
        keys:
          $ref: "#/components/schemas/PushSubscriptionKeys"
        userAgent:
          type: string
          description: Optional User-Agent string for diagnostics
      required:
        - endpoint
        - keys

    PushUnsubscribeRequest:
      type: object
      properties:
        endpoint:
          type: string
      required:
        - endpoint

    PushSubscribeResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok

    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Candle:
      type: object
      properties:
        date:
          type: string
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
      required: [date, open, high, low, close, volume]

    PriceHistory:
      type: object
      properties:
        symbol:
          type: string
        candles:
          type: array
          items:
            $ref: "#/components/schemas/Candle"
        currentPrice:
          type: number
          description: Live OANDA spot price (also used to align candles)
        lastUpdated:
          type: string
      required: [symbol, candles, currentPrice, lastUpdated]

    PriceLevel:
      type: object
      properties:
        label:
          type: string
          description: e.g. "R3", "S1", "Fib 61.8%", "Swing High"
        price:
          type: number
        type:
          type: string
          enum: [resistance, support, pivot]
      required: [label, price, type]

    TradeZone:
      type: object
      properties:
        low:
          type: number
        high:
          type: number
        label:
          type: string
      required: [low, high, label]

    LevelsData:
      type: object
      properties:
        symbol:
          type: string
        currentPrice:
          type: number
        priceChange:
          type: number
        priceChangePct:
          type: number
        signal:
          type: string
          enum: [BUY, SELL, WAIT]
          description: Single clear trade signal
        signalReason:
          type: string
          description: Plain-language explanation of why this signal was generated
        tradeState:
          type: string
          enum: [WAIT, PENDING, FILLED_PROFIT, FILLED_DRAWDOWN, FILLED_TP1, FILLED_TP2, FILLED_SL]
          description: Machine-readable trade lifecycle state. WAIT = no setup. PENDING = limit staged but not yet tagged. FILLED_DRAWDOWN/PROFIT = position open between entry and SL/TP1. FILLED_TP1/TP2/SL = price has reached that level (TP2/SL trigger invalidation on the next tick). Always prefer this over parsing signalReason text.
        entryPrice:
          type: number
          description: Ideal entry price
        stopLoss:
          type: number
          description: Recommended stop loss
        takeProfit1:
          type: number
          description: First take profit target
        takeProfit2:
          type: number
          description: Second take profit target
        riskRewardRatio:
          type: number
          description: Risk/reward ratio for the trade
        buyZone:
          $ref: "#/components/schemas/TradeZone"
        sellZone:
          $ref: "#/components/schemas/TradeZone"
        levels:
          type: array
          items:
            $ref: "#/components/schemas/PriceLevel"
          description: All key support and resistance levels
        pivot:
          type: number
          description: Daily pivot point
        trend:
          type: string
          enum: [UPTREND, DOWNTREND, RANGING]
          description: Current market structure
        trendStrength:
          type: number
          description: 0-100 strength score
        lastUpdated:
          type: string
        positionSizing:
          $ref: "#/components/schemas/PositionSizing"
      required:
        - symbol
        - currentPrice
        - priceChange
        - priceChangePct
        - signal
        - signalReason
        - tradeState
        - entryPrice
        - stopLoss
        - takeProfit1
        - takeProfit2
        - riskRewardRatio
        - buyZone
        - sellZone
        - levels
        - pivot
        - trend
        - trendStrength
        - lastUpdated

    PositionSizing:
      type: object
      description: Suggested position size for the trader's account. The `venue` field decides which downstream block is authoritative — JUP (crypto perps) renders `achievable`, MT5 (forex/metals) renders `mt5`.
      properties:
        venue:
          type: string
          enum: [JUP, MT5]
          description: Which trading venue this symbol uses. JUP = Jupiter perps (BTC/ETH, collateral × leverage). MT5 = MetaTrader 5 (forex/metals, lot-based sizing).
        accountSize:
          type: number
          description: Account size in USD used for sizing
        riskAmount:
          type: number
          description: Dollar amount risked per trade if stop is hit (JUP venue's risk-budget; on MT5 venue this is the *would-be* budget, not the actual lot-based risk — read mt5.pnlAtSL for the actual loss)
        riskPct:
          type: number
          description: Risk percent of account (e.g. 1.0 = 1%)
        positionSize:
          type: number
          description: Position size in base units (coins, oz, base currency)
        positionSizeUnit:
          type: string
          description: Unit label for positionSize (BTC, ETH, oz, EUR, etc.)
        notional:
          type: number
          description: Approximate USD notional value of the position
        leverage:
          type: number
          description: Suggested minimum leverage multiplier (perps only)
        leverageNote:
          type: string
          description: Warning text when required leverage is high
        lots:
          type: object
          description: Lot breakdown for forex / metals
          properties:
            standard:
              type: number
            mini:
              type: number
            micro:
              type: number
          required: [standard, mini, micro]
        achievable:
          $ref: "#/components/schemas/AchievablePosition"
        mt5:
          $ref: "#/components/schemas/MT5Sizing"
      required:
        - venue
        - accountSize
        - riskAmount
        - riskPct
        - positionSize
        - positionSizeUnit
        - notional

    MT5Sizing:
      type: object
      description: MetaTrader 5 lot-based sizing for forex and metals. The trader chooses a fixed lot size (default 0.01 = 1 micro lot); the dollar P&L at SL/TP1/TP2 is derived from the pair's contract size and quote-currency conversion. For X/USD pairs this is exact; for USDJPY it uses the live entry price; for GBPJPY it approximates via an assumed USDJPY rate.
      properties:
        lots:
          type: number
          description: Lot size the trader will place (e.g. 0.01 = 1 micro lot, 0.10 = 1 mini lot, 1.00 = 1 standard lot)
        contractSize:
          type: number
          description: Units of base asset per 1.0 standard lot (100,000 for forex, 100 oz for XAU, 5,000 oz for XAG)
        positionSize:
          type: number
          description: Total base units this trade controls (lots × contractSize)
        positionSizeUnit:
          type: string
          description: Unit label for positionSize (EUR, GBP, AUD, USD, oz)
        notional:
          type: number
          description: Approximate USD notional value of the position
        pnlAtSL:
          type: number
          description: Signed dollar P&L if stop loss hits (always negative)
        pnlAtTP1:
          type: number
          description: Signed dollar P&L if TP1 hits (positive)
        pnlAtTP2:
          type: number
          description: Signed dollar P&L if TP2 hits (positive)
        riskPctOfAccount:
          type: number
          description: "Absolute pnlAtSL as a percent of the account size — useful for sanity-checking that 0.01 lots isn't over-leveraging the account on a high-pip-value pair like XAGUSD"
        recommendedLots:
          type: number
          description: "Lot size that would risk approximately the configured riskPct of the account on SL hit. Rounded down to 0.01 increments and clamped to a sane min/max. Use this to size a trade that's actually worth taking on small accounts where 0.01 lots yields cents."
        recommendedTargetRiskPct:
          type: number
          description: "The riskPct used to derive recommendedLots (mirrors the request's riskPct query param)."
      required:
        - lots
        - contractSize
        - positionSize
        - positionSizeUnit
        - notional
        - pnlAtSL
        - pnlAtTP1
        - pnlAtTP2
        - riskPctOfAccount
        - recommendedLots
        - recommendedTargetRiskPct

    AchievablePosition:
      type: object
      description: Position scaled to honour exchange minimums (e.g. Jupiter $10 minimum collateral). Shows what the trader can ACTUALLY take and the real dollar P&L at each level.
      properties:
        positionSize:
          type: number
          description: Actual position size in base units after applying exchange constraints
        notional:
          type: number
          description: Actual notional value in USD
        collateral:
          type: number
          description: Recommended collateral to deposit (USD). At least minCollateral.
        leverage:
          type: number
          description: Leverage to set on the exchange so notional = collateral × leverage
        actualRiskAmount:
          type: number
          description: Actual dollar loss if SL is hit (may differ from intended riskAmount when minimum forces over-sizing)
        actualRiskPct:
          type: number
          description: actualRiskAmount as a percent of account
        pnlAtSL:
          type: number
          description: Signed dollar P&L if stop loss hits (always negative)
        pnlAtTP1:
          type: number
          description: Signed dollar P&L if TP1 hits (positive)
        pnlAtTP2:
          type: number
          description: Signed dollar P&L if TP2 hits (positive)
        belowMinimum:
          type: boolean
          description: True when ideal notional was below the exchange minimum and position was forced over-sized
        warning:
          type: string
          description: Human-readable warning when constraints affected the trade
      required:
        - positionSize
        - notional
        - collateral
        - leverage
        - actualRiskAmount
        - actualRiskPct
        - pnlAtSL
        - pnlAtTP1
        - pnlAtTP2
        - belowMinimum

    ActiveSignalEntry:
      type: object
      description: One active BUY/SELL signal in the overview. Wraps a LevelsData payload with the timeframe so the client can route on click.
      properties:
        symbol:
          type: string
          enum: ["XAGUSD", "XAUUSD", "EURUSD", "GBPUSD", "AUDUSD", "USDJPY", "GBPJPY", "BTCUSD", "ETHUSD"]
        timeframe:
          type: string
          enum: ["15m", "30m", "1h", "1d"]
        levels:
          $ref: "#/components/schemas/LevelsData"
      required: [symbol, timeframe, levels]

    ActiveSignalsCoverage:
      type: object
      description: Per-request data-feed health. Lets the UI distinguish "no signals" (succeeded == total, signals empty) from "data feed degraded" (failed > 0).
      properties:
        total:
          type: integer
          description: Total symbol × timeframe combos attempted
        succeeded:
          type: integer
          description: Combos that returned a usable result (BUY, SELL, or WAIT)
        failed:
          type: integer
          description: Combos that errored or had insufficient data
        failedSymbols:
          type: array
          items:
            type: string
          description: Distinct symbols with at least one failed combo
      required: [total, succeeded, failed, failedSymbols]

    ActiveSignalsResponse:
      type: object
      properties:
        signals:
          type: array
          items:
            $ref: "#/components/schemas/ActiveSignalEntry"
          description: Every currently-active BUY/SELL signal. Empty array means no actionable signals right now.
        coverage:
          $ref: "#/components/schemas/ActiveSignalsCoverage"
        lastUpdated:
          type: string
      required: [signals, coverage, lastUpdated]

    BacktestTrade:
      type: object
      properties:
        entryDate:
          type: string
        direction:
          type: string
          enum: [BUY, SELL]
        entry:
          type: number
        stopLoss:
          type: number
        takeProfit1:
          type: number
        takeProfit2:
          type: number
        exitDate:
          type: string
        exitPrice:
          type: number
        outcome:
          type: string
          enum: [TP1, TP2, SL, EXPIRED]
        rMultiple:
          type: number
          description: Profit/loss in units of risk (R)
        barsHeld:
          type: integer
      required:
        - entryDate
        - direction
        - entry
        - stopLoss
        - takeProfit1
        - takeProfit2
        - exitDate
        - exitPrice
        - outcome
        - rMultiple
        - barsHeld

    BacktestResult:
      type: object
      properties:
        symbol:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        totalBars:
          type: integer
        totalTrades:
          type: integer
        winningTrades:
          type: integer
        losingTrades:
          type: integer
        winRate:
          type: number
          description: Percentage of trades that hit at least TP1
        totalReturnR:
          type: number
          description: Cumulative R-multiples
        avgReturnR:
          type: number
          description: Average R per trade
        profitFactor:
          type: number
          description: Gross profit / gross loss
        maxDrawdownR:
          type: number
          description: Largest peak-to-trough drawdown in R
        avgBarsHeld:
          type: number
        buyTrades:
          type: integer
        sellTrades:
          type: integer
        buyWinRate:
          type: number
        sellWinRate:
          type: number
        tp1Hits:
          type: integer
        tp2Hits:
          type: integer
        slHits:
          type: integer
        expiredHits:
          type: integer
        trades:
          type: array
          items:
            $ref: "#/components/schemas/BacktestTrade"
          description: Most recent trades (capped to last 50 for response size)
        lastUpdated:
          type: string
      required:
        - symbol
        - startDate
        - endDate
        - totalBars
        - totalTrades
        - winningTrades
        - losingTrades
        - winRate
        - totalReturnR
        - avgReturnR
        - profitFactor
        - maxDrawdownR
        - avgBarsHeld
        - buyTrades
        - sellTrades
        - buyWinRate
        - sellWinRate
        - tp1Hits
        - tp2Hits
        - slHits
        - expiredHits
        - trades
        - lastUpdated