# lib/api-zod/src/generated/types/pushSubscriptionKeys.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XAGUSD Screener API
 * OpenAPI spec version: 0.2.0
 */

export interface PushSubscriptionKeys {
  /** Browser-supplied P-256 ECDH public key (base64url) */
  p256dh: string;
  /** Browser-supplied auth secret (base64url) */
  auth: string;
}