# lib/api-zod/src/generated/types/pushSubscriptionPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XAGUSD Screener API
 * OpenAPI spec version: 0.2.0
 */
import type { PushSubscriptionKeys } from "./pushSubscriptionKeys";

export interface PushSubscriptionPayload {
  /** Push service endpoint URL (FCM / APNs gateway) */
  endpoint: string;
  keys: PushSubscriptionKeys;
  /** Optional User-Agent string for diagnostics */
  userAgent?: string;
}