# lib/api-zod/src/generated/types/vapidPublicKey.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XAGUSD Screener API
 * OpenAPI spec version: 0.2.0
 */

export interface VapidPublicKey {
  /** Base64URL-encoded VAPID public key suitable for PushManager.subscribe applicationServerKey */
  publicKey: string;
}